209 documentation and comments local branch

+++ API/Services/HotelService.cs
@@ -12,12 +12,20 @@ namespace API.Services
     {
         private readonly AppDBContext _context;
         private readonly HotelMapping _hotelMapping = new();
+        /// <summary>
+        /// Constructor for the HotelService class
+        /// </summary>
+        /// <param name="context">Database context</param>
         public HotelService(AppDBContext context)
         {
             _context = context;
         }
 
-        //GET
+        /// <summary>
+        /// Retrieves all the hotels fom the database
+        /// </summary>
+        /// <returns>The hotels</returns>
+        /// <exception cref="ArgumentException"></exception>
         public async Task<IEnumerable<HotelDto>> GetHotel()
         {
             var hotels = await _context.Hotels.ToListAsync()
@@ -28,7 +36,12 @@ namespace API.Services
             .ToList();
         }
 
-        //GET {Id}
+        /// <summary>
+        /// View one specific hotel by id
+        /// </summary>
+        /// <param name="id">unique identifier for the hotel</param>
+        /// <returns>The specified hotel</returns>
+        /// <exception cref="ArgumentException"></exception>
         public async Task<HotelDto> GetHotelById(int id)
         {
             if (id == 0)
@@ -39,42 +52,34 @@ namespace API.Services
 
             return _hotelMapping.ToHotelGETdto(hotel);
         }
-
-        //Use type Hotel instead of HotelDto, as we want the new hotel into the DB
-        public async Task<Hotel> PostHotel(HotelDto hotelCreateDto)
+        
+        /// <summary>
+        /// Creates a new hotel
+        /// </summary>
+        /// <param name="hotelCreateDto">Details for the new hotel</param>
+        /// <returns>The new created hotel</returns>
+        /// <exception cref="ArgumentException"></exception>
+        public async Task<HotelDto> PostHotel(HotelDto hotelCreateDto)
         {
             //Check if hotel.name already exists in our database
             if (await _context.Hotels.AnyAsync(h => h.HotelName == hotelCreateDto.HotelName))
                 throw new ArgumentException($"Hotel already exist: {hotelCreateDto.HotelName}");
-            
-            var newHotel = new Hotel
-            {
-                HotelName = hotelCreateDto.HotelName,
-                CityName = hotelCreateDto.CityName,
-                Address = hotelCreateDto.Address,
-                Description = hotelCreateDto.Description,
-                CreatedAt = DateTime.UtcNow.AddHours(2),
-                UpdatedAt = DateTime.UtcNow.AddHours(2),
-                Email = hotelCreateDto.Email,
-                Phone = hotelCreateDto.Phone,
-                WeekdayTime = hotelCreateDto.WeekdayTime,
-                SaturdayTime = hotelCreateDto.SaturdayTime,
-                HolidaysTime = hotelCreateDto.HolidaysTime
-            };
+
+            var newHotel = _hotelMapping.ToHotelPOSTDto(hotelCreateDto);
 
             _context.Hotels.Add(newHotel);
             await _context.SaveChangesAsync();
 
-            var createdHotel = _context.Hotels.FirstOrDefault(h => h.HotelName == newHotel.HotelName)
-                ?? throw new Exception("Something went wrong saving the hotel to the database!");
 
-            //Return new Hotel into db
-            return createdHotel;
+            return _hotelMapping.ToHotelGETdto(newHotel);
         }
 
         /// <summary>
-        /// Returns HotelDto, as we want the "tailored" version
+        /// Updates an hotel
         /// </summary>
+        /// <param name="updatedHotel">The details of hotel we want updated</param>
+        /// <returns>The new updated hotel</returns>
+        /// <exception cref="ArgumentException"></exception>
         public async Task<HotelDto?> PutHotel(HotelDto updatedHotel)
         {
             var currentHotel = await _context.Hotels.FindAsync(updatedHotel.Id)
@@ -83,8 +88,6 @@ namespace API.Services
             _hotelMapping.TohotelPUTDto(currentHotel, updatedHotel);
 
             await _context.SaveChangesAsync();
-
-            // Return the updated hotel as DTO
             return new HotelDto
             {
                 Id = currentHotel.Id,
@@ -100,13 +103,20 @@ namespace API.Services
             };
         }
 
+        /// <summary>
+        /// Deletes an hotel
+        /// </summary>
+        /// <param name="id">Id of the hotel we want deleted</param>
+        /// <returns>True, if the hotel was deleted</returns>
         public async Task<bool> DeleteHotel(int id)
         {
+            if (id == 0)
+                throw new ArgumentException("Id can't be 0");
+
             var hotel = await _context.Hotels.FindAsync(id);
             if (hotel == null)
-            {
-                return false;
-            }
+                throw new ArgumentException($"Cound't find hotel with the Id: {id}");
+
             _context.Hotels.Remove(hotel);
             await _context.SaveChangesAsync();
